+++ RELEASES.md
@@ -1,4 +1,4 @@
-# Release v0.5.6b 2026-02-10
+# Release v0.5.6b 2026-02-11
 
 ## New Features
 
@@ -6,6 +6,7 @@
 2. modify builds to produce valgrind compatible code (debug only) (v0.5.6b 2026-02-10)
 3. ability to restart server to listen on new port while running by changing port or interface via
   `CONFIG SET server_binding` or `server_port`
+4. allow `START` function to restart server on the fly (v0.5.6b 2026-02-11)
 
 # Release v0.5.5b 2026-02-06